requirements/package.txt
six
djangorestframework>=3.7,<4.0
drf-yasg>=1.4,<2.0
djangorestframework-jwt>=1.11,<2.0
django-allauth>=0.36,<1.0
